+++ app/src/main/java/com/github/se/project/MainActivity.kt
@@ -9,15 +9,18 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navArgument
 import com.github.se.project.model.authentification.AuthenticationViewModel
 import com.github.se.project.model.lesson.LessonViewModel
 import com.github.se.project.model.profile.ListProfilesViewModel
 import com.github.se.project.ui.authentification.SignInScreen
 import com.github.se.project.ui.lesson.AddLessonScreen
+import com.github.se.project.ui.lesson.EditRequestedLessonScreen
 import com.github.se.project.ui.navigation.NavigationActions
 import com.github.se.project.ui.navigation.Route
 import com.github.se.project.ui.navigation.Screen
@@ -31,6 +34,7 @@ import com.github.se.project.ui.theme.SampleAppTheme
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
+    // enableEdgeToEdge()
     setContent {
       SampleAppTheme { Surface(modifier = Modifier.fillMaxSize()) { PocketTutorApp() } }
     }
@@ -135,10 +139,15 @@ fun PocketTutorApp() {
       composable(Screen.ADD_LESSON) {
         AddLessonScreen(navigationActions, listProfilesViewModel, lessonViewModel)
       }
-      // composable(Screen.EDIT_LESSON) {
-      //  EditLessonScreen(listProfilesViewModel, lessonViewModel, navigationActions)
-      // }
+      composable(
+          Screen.EDIT_REQUESTED_LESSON + "/{Lesson ID}",
+          arguments = listOf(navArgument("Lesson ID") { type = NavType.StringType })) { entry ->
+            val lessonId = entry.arguments?.getString("Lesson ID")!!
+            EditRequestedLessonScreen(
+                lessonId, navigationActions, listProfilesViewModel, lessonViewModel)
+          }
 
+      // composable(Screen.EDIT_SCHEDULED_LESSON)
     }
   }
 }
