+++ app/src/main/java/com/github/se/project/model/lesson/Lesson.kt
@@ -4,7 +4,7 @@ import com.github.se.project.model.profile.Language
 import com.github.se.project.model.profile.Subject
 
 data class Lesson(
-    val id: String = "", // Unique identifier for the lesson
+    var id: String = "", // Unique identifier for the lesson
     val title: String = "", // Title of the lesson
     val description: String = "", // Description of the lesson
     val subject: Subject = Subject.NONE,
@@ -14,6 +14,6 @@ data class Lesson(
     val minPrice: Double = 0.0, // Price for the lesson
     val maxPrice: Double = 0.0, // Price for the lesson
     val price: Double = 0.0, // Price for the lesson
-    val timeSlot: String = "", // Time slot for the lesson (e.g., "2024-10-10T10:00:00")
+    val timeSlot: String = "", // Time slot for the lesson (e.g., "30/10/2024T10:00:00")
     val status: LessonStatus = LessonStatus.PENDING, // Status of the lesson
 )
