+++ app/src/main/java/com/github/se/project/model/lesson/LessonViewModel.kt
@@ -3,13 +3,11 @@ package com.github.se.project.model.lesson
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewModelScope
 import com.google.firebase.Firebase
 import com.google.firebase.firestore.firestore
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
 
 /**
  * ViewModel for managing lessons and interacting with the LessonRepository. Handles the retrieval,
@@ -134,10 +132,16 @@ open class LessonViewModel(private val repository: LessonRepository) : ViewModel
         })
   }
 
-    // Method to update a lesson by passing it to the repository
-    fun updateLesson(lesson: Lesson, onSuccess: () -> Unit, onFailure: (Exception) -> Unit) {
-        viewModelScope.launch {
-            repository.updateLesson(lesson, onSuccess, onFailure)
-        }
-    }
+  // Method to update a lesson by passing it to the repository
+  fun updateLesson(lesson: Lesson, onComplete: () -> Unit) {
+    repository.updateLesson(
+        lesson,
+        onSuccess = {
+          onComplete() // Call the provided callback on success
+        },
+        onFailure = {
+          Log.e("LessonViewModel", "Error adding lesson: $lesson", it)
+          onComplete() // Call the callback even if there's a failure
+        })
+  }
 }
