+++ app/src/main/java/com/github/se/project/ui/overview/Overview.kt
@@ -53,6 +53,10 @@ fun HomeScreen(
         else -> LIST_TOP_LEVEL_DESTINATIONS_STUDENT
       }
 
+  val onLessonClick = { lesson: Lesson ->
+    navigationActions.navigateTo(Screen.EDIT_REQUESTED_LESSON + "/${lesson.id}")
+  }
+
   Scaffold(
       topBar = {
         Row(
@@ -102,6 +106,7 @@ fun HomeScreen(
                         statusFilter = LessonStatus.PENDING,
                         isTutor = true,
                         maxHeight = 340.dp,
+                        onClick = onLessonClick,
                         modifier = Modifier.testTag("pendingLessonsSection"))
                     Spacer(modifier = Modifier.height(16.dp))
 
@@ -110,6 +115,7 @@ fun HomeScreen(
                         lessons = lessons,
                         statusFilter = LessonStatus.CONFIRMED,
                         isTutor = true,
+                        onClick = onLessonClick,
                         modifier = Modifier.testTag("confirmedLessonsSection"))
                   }
 
@@ -121,6 +127,7 @@ fun HomeScreen(
                         statusFilter = LessonStatus.REQUESTED,
                         isTutor = false,
                         maxHeight = 340.dp,
+                        onClick = onLessonClick,
                         modifier = Modifier.testTag("requestedLessonsSection"))
                     Spacer(modifier = Modifier.height(16.dp))
 
@@ -129,6 +136,7 @@ fun HomeScreen(
                         lessons = lessons,
                         statusFilter = LessonStatus.CONFIRMED,
                         isTutor = false,
+                        onClick = onLessonClick,
                         modifier = Modifier.testTag("confirmedLessonsStudentSection"))
                   }
                 }
@@ -150,7 +158,8 @@ fun ExpandableLessonSection(
     statusFilter: LessonStatus,
     isTutor: Boolean,
     maxHeight: Dp? = null, // Optional parameter to limit the section's height
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
+    onClick: (Lesson) -> Unit,
 ) {
   var expanded by remember { mutableStateOf(true) }
 
@@ -183,6 +192,7 @@ fun ExpandableLessonSection(
             lessons = lessons,
             statusFilter = statusFilter,
             isTutor = isTutor,
+            onClick = onClick,
             modifier = Modifier.testTag("${sectionTitle}Lessons"))
       }
     }
